+++ dogm/include/dogm.h
@@ -35,57 +35,57 @@ struct GridCell
 {
 	int start_idx;
 	int end_idx;
-	double new_born_occ_mass;
-	double pers_occ_mass;
-	double free_mass;
-	double occ_mass;
-	double mu_A;
-	double mu_UA;
+	float new_born_occ_mass;
+	float pers_occ_mass;
+	float free_mass;
+	float occ_mass;
+	float mu_A;
+	float mu_UA;
 
-	double w_A;
-	double w_UA;
+	float w_A;
+	float w_UA;
 
-	double mean_x_vel;
-	double mean_y_vel;
-	double var_x_vel;
-	double var_y_vel;
-	double covar_xy_vel;
+	float mean_x_vel;
+	float mean_y_vel;
+	float var_x_vel;
+	float var_y_vel;
+	float covar_xy_vel;
 
 	int2 pos;
 };
 
 struct MeasurementCell
 {
-	double free_mass;
-	double occ_mass;
-	double likelihood;
-	double p_A;
+	float free_mass;
+	float occ_mass;
+	float likelihood;
+	float p_A;
 };
 
 struct Particle
 {
 	int grid_cell_idx;
-	double weight;
+	float weight;
 	bool associated;
 	glm::vec4 state;
 };
 
 struct GridParams
 {
-	double size;
-	double resolution;
+	float size;
+	float resolution;
 	int particle_count;
 	int new_born_particle_count;
-	double persistence_prob;
-	double process_noise_position;
-	double process_noise_velocity;
-	double birth_prob;
+	float persistence_prob;
+	float process_noise_position;
+	float process_noise_velocity;
+	float birth_prob;
 };
 
 struct LaserSensorParams
 {
-	double max_range;
-	double fov;
+	float max_range;
+	float fov;
 };
 
 class DOGM
@@ -94,8 +94,8 @@ public:
 	DOGM(const GridParams& params, const LaserSensorParams& laser_params);
 	~DOGM();
 
-	void updateMeasurementGrid(double* measurements, int num_measurements);
-	void updateParticleFilter(double dt);
+	void updateMeasurementGrid(float* measurements, int num_measurements);
+	void updateParticleFilter(float dt);
 
 	int getGridSize() const { return grid_size; }
 
@@ -103,7 +103,7 @@ private:
 	void initialize();
 
 public:
-	void particlePrediction(double dt);
+	void particlePrediction(float dt);
 	void particleAssignment();
 	void gridCellOccupancyUpdate();
 	void updatePersistentParticles();
@@ -126,9 +126,9 @@ public:
 	MeasurementCell* polar_meas_cell_array;
 	MeasurementCell* meas_cell_array;
 
-	double* weight_array;
-	double* birth_weight_array;
-	double* born_masses_array;
+	float* weight_array;
+	float* birth_weight_array;
+	float* born_masses_array;
 
 	int grid_size;
