+++ dogm/include/kernel/init_new_particles.h
@@ -26,11 +26,11 @@ SOFTWARE.
 #include "dogm.h"
 #include <device_launch_parameters.h>
 
-void normalize_particle_orders(double* particle_orders_array_accum, int particle_orders_count, int v_B);
+void normalize_particle_orders(float* particle_orders_array_accum, int particle_orders_count, int v_B);
 
 __global__ void initNewParticlesKernel1(Particle* particle_array, GridCell* grid_cell_array, MeasurementCell* meas_cell_array,
-	double* weight_array, double* born_masses_array, Particle* birth_particle_array, double* particle_orders_array_accum, int cell_count);
+	float* weight_array, float* born_masses_array, Particle* birth_particle_array, float* particle_orders_array_accum, int cell_count);
 
 __global__ void initNewParticlesKernel2(Particle* birth_particle_array, GridCell* grid_cell_array, int grid_size, int particle_count);
 
-__global__ void copyBirthWeightKernel(Particle* birth_particle_array, double* birth_weight_array, int particle_count);
+__global__ void copyBirthWeightKernel(Particle* birth_particle_array, float* birth_weight_array, int particle_count);
