+++ dogm/include/kernel/resampling.h
@@ -28,8 +28,8 @@ SOFTWARE.
 #include <thrust/device_vector.h>
 #include <thrust/host_vector.h>
 
-void calc_resampled_indeces(thrust::device_vector<double>& weight_accum, thrust::device_vector<int>& rand_array,
+void calc_resampled_indeces(thrust::device_vector<float>& weight_accum, thrust::device_vector<int>& rand_array,
 	thrust::device_vector<int>& indices);
 
 __global__ void resamplingKernel(Particle* particle_array, Particle* particle_array_next, Particle* birth_particle_array,
-	int* idx_array_resampled, double joint_max, int particle_count);
+	int* idx_array_resampled, float joint_max, int particle_count);
