+++ dogm/src/kernel/mass_update.cu
@@ -30,45 +30,45 @@ SOFTWARE.
 
 #include <thrust/execution_policy.h>
 
-__device__ double predict_free_mass(GridCell& grid_cell, double m_occ_pred, double alpha = 0.9)
+__device__ float predict_free_mass(GridCell& grid_cell, float m_occ_pred, float alpha = 0.9)
 {
-	double m_free_pred = min(alpha * grid_cell.free_mass, 1.0 - m_occ_pred);
+	float m_free_pred = min(alpha * grid_cell.free_mass, 1.0 - m_occ_pred);
 
 	// limit free mass
-	double sum = m_free_pred + m_occ_pred;
+	float sum = m_free_pred + m_occ_pred;
 	if (sum > 1.0)
 	{
-		double diff = sum - 1.0;
+		float diff = sum - 1.0;
 		m_free_pred -= diff;
 	}
 
 	return m_free_pred;
 }
 
-__device__ double update_o(double m_occ_pred, double m_free_pred, const MeasurementCell& meas_cell)
+__device__ float update_o(float m_occ_pred, float m_free_pred, const MeasurementCell& meas_cell)
 {
-	double unknown_pred = 1.0 - m_occ_pred - m_free_pred;
-	double meas_cell_unknown = 1.0 - meas_cell.free_mass - meas_cell.occ_mass;
-	double K = m_free_pred * meas_cell.occ_mass + m_occ_pred * meas_cell.free_mass;
+	float unknown_pred = 1.0 - m_occ_pred - m_free_pred;
+	float meas_cell_unknown = 1.0 - meas_cell.free_mass - meas_cell.occ_mass;
+	float K = m_free_pred * meas_cell.occ_mass + m_occ_pred * meas_cell.free_mass;
 
 	return (m_occ_pred * meas_cell_unknown + unknown_pred * meas_cell.occ_mass + m_occ_pred * meas_cell.occ_mass) / (1.0 - K);
 }
 
-__device__ double update_f(double m_occ_pred, double m_free_pred, const MeasurementCell& meas_cell)
+__device__ float update_f(float m_occ_pred, float m_free_pred, const MeasurementCell& meas_cell)
 {
-	double unknown_pred = 1.0 - m_occ_pred - m_free_pred;
-	double meas_cell_unknown = 1.0 - meas_cell.free_mass - meas_cell.occ_mass;
-	double K = m_free_pred * meas_cell.occ_mass + m_occ_pred * meas_cell.free_mass;
+	float unknown_pred = 1.0 - m_occ_pred - m_free_pred;
+	float meas_cell_unknown = 1.0 - meas_cell.free_mass - meas_cell.occ_mass;
+	float K = m_free_pred * meas_cell.occ_mass + m_occ_pred * meas_cell.free_mass;
 
 	return (m_free_pred * meas_cell_unknown + unknown_pred * meas_cell.free_mass + m_free_pred * meas_cell.free_mass) / (1.0 - K);
 }
 
-__device__ double separate_newborn_part(double m_occ_pred, double m_occ_up, double p_B)
+__device__ float separate_newborn_part(float m_occ_pred, float m_occ_up, float p_B)
 {
 	return (m_occ_up * p_B * (1.0 - m_occ_pred)) / (m_occ_pred + p_B * (1.0 - m_occ_pred));
 }
 
-__device__ void store_values(double rho_b, double rho_p, double m_free_up, double m_occ_up, GridCell* grid_cell_array, int i)
+__device__ void store_values(float rho_b, float rho_p, float m_free_up, float m_occ_up, GridCell* grid_cell_array, int i)
 {
 	grid_cell_array[i].pers_occ_mass = rho_p;
 	grid_cell_array[i].new_born_occ_mass = rho_b;
@@ -76,9 +76,9 @@ __device__ void store_values(double rho_b, double rho_p, double m_free_up, doubl
 	grid_cell_array[i].occ_mass = m_occ_up;
 }
 
-__device__ void normalize_to_pS(Particle* particle_array, double* weight_array, double p_S, int start_idx, int end_idx)
+__device__ void normalize_to_pS(Particle* particle_array, float* weight_array, float p_S, int start_idx, int end_idx)
 {
-	double sum = 0.0f;
+	float sum = 0.0f;
 	for (int i = start_idx; i < end_idx + 1; i++)
 	{
 		sum += weight_array[i];
@@ -91,15 +91,15 @@ __device__ void normalize_to_pS(Particle* particle_array, double* weight_array, 
 	}
 }
 
-__global__ void gridCellPredictionUpdateKernel(GridCell* grid_cell_array, Particle* particle_array, double* weight_array,
-	double* weight_array_accum, MeasurementCell* meas_cell_array, double* born_masses_array, double p_B, double p_S, int cell_count)
+__global__ void gridCellPredictionUpdateKernel(GridCell* grid_cell_array, Particle* particle_array, float* weight_array,
+	float* weight_array_accum, MeasurementCell* meas_cell_array, float* born_masses_array, float p_B, float p_S, int cell_count)
 {
 	for (int i = blockIdx.x * blockDim.x + threadIdx.x; i < cell_count; i += blockDim.x * gridDim.x)
 	{
 		int start_idx = grid_cell_array[i].start_idx;
 		int end_idx = grid_cell_array[i].end_idx;
 
-		double m_occ_pred = subtract(weight_array_accum, start_idx, end_idx);
+		float m_occ_pred = subtract(weight_array_accum, start_idx, end_idx);
 
 		if (m_occ_pred > p_S)
 		{
@@ -107,11 +107,11 @@ __global__ void gridCellPredictionUpdateKernel(GridCell* grid_cell_array, Partic
 			normalize_to_pS(particle_array, weight_array, p_S, start_idx, end_idx);
 		}
 
-		double m_free_pred = predict_free_mass(grid_cell_array[i], m_occ_pred);
-		double m_occ_up = update_o(m_occ_pred, m_free_pred, meas_cell_array[i]);
-		double m_free_up = update_f(m_occ_pred, m_free_pred, meas_cell_array[i]);
-		double rho_b = separate_newborn_part(m_occ_pred, m_occ_up, p_B);
-		double rho_p = m_occ_up - rho_b;
+		float m_free_pred = predict_free_mass(grid_cell_array[i], m_occ_pred);
+		float m_occ_up = update_o(m_occ_pred, m_free_pred, meas_cell_array[i]);
+		float m_free_up = update_f(m_occ_pred, m_free_pred, meas_cell_array[i]);
+		float rho_b = separate_newborn_part(m_occ_pred, m_occ_up, p_B);
+		float rho_p = m_occ_up - rho_b;
 		born_masses_array[i] = rho_b;
 		store_values(rho_b, rho_p, m_free_up, m_occ_up, grid_cell_array, i);
 	}
